Bert_Pipeline/requirements.txt
absl-py==1.0.0
astor==0.8.1
bert-tensorflow==1.0.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
dataclasses==0.8
et-xmlfile==1.1.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
numpy==1.19.5
openpyxl==3.0.9
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.19.4
python-dateutil==2.8.2
pytz==2021.3
regex==2022.1.18
requests==2.27.1
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-serving-api==1.15.0
termcolor==1.1.0
tokenization==1.0.7
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.6.0
Flask==2.0.2
